Move AlignmentInfo into ibm_model
Because AlignmentInfo is a reusable data structure for IBM Models 3 and
higher.
